app/model-services/requirements.txt
# Model Services Base Requirements
# Individual services can add their own specific requirements

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
pillow==10.1.0
python-dotenv==1.0.0

# Database access (if needed)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# For AI Models (Hybrid AI Integration)
transformers>=4.30.0  # For BioGPT and other language models
torch>=2.0.0  # PyTorch for AI models
torchvision>=0.15.0  # For image models

# For ML Models (Hybrid AI Integration)
lightgbm>=4.0.0  # For vital risk prediction
